Refactor tests to packages
