+++ src/test/java/secondWeek/Ordersystem/db/OrderDbTester.java
@@ -1,4 +1,4 @@
-package secondWeek.Ordersystem;
+package secondWeek.Ordersystem.db;
 
 import org.junit.jupiter.api.Test;
 import secondWeek.Ordersystem.db.OrderDb;
