+++ src/test/java/secondWeek/Ordersystem/services/OrderServiceTest.java
@@ -1,4 +1,4 @@
-package secondWeek.Ordersystem;
+package secondWeek.Ordersystem.services;
 
 import org.junit.jupiter.api.Test;
 import secondWeek.Ordersystem.model.Order;
